ci: use one test summary status
Collapse unit and E2E status reporting into a single Summary job, and document the deployment tarball wait and dependency-free test glob implementation.

Made-with: Cursor
