+++ .github/workflows/test.yml
@@ -44,24 +44,19 @@ jobs:
       # GitHub commit statuses are upserted by `(sha, context)`, so a
       # mismatch would create a second row instead of overwriting this
       # pending one with the final result.
-      - name: Post pending Summary statuses to PR commit
+      - name: Post pending Summary status to PR commit
         uses: actions/github-script@v7
         with:
           script: |
-            for (const { context: statusContext, description } of [
-              { context: 'Summary', description: 'Unit tests in progress' },
-              { context: 'Summary (e2e)', description: 'E2E tests in progress' },
-            ]) {
-              await github.rest.repos.createCommitStatus({
-                owner: context.repo.owner,
-                repo: context.repo.repo,
-                sha: '${{ github.event.pull_request.head.sha }}',
-                state: 'pending',
-                context: statusContext,
-                target_url: `${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}`,
-                description,
-              });
-            }
+            await github.rest.repos.createCommitStatus({
+              owner: context.repo.owner,
+              repo: context.repo.repo,
+              sha: '${{ github.event.pull_request.head.sha }}',
+              state: 'pending',
+              context: 'Summary',
+              target_url: `${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}`,
+              description: 'Tests in progress',
+            });
       - name: Resolve PR context
         id: pr-context
         env:
@@ -384,6 +379,14 @@ jobs:
         run: |
           set -euo pipefail
 
+          # E2E needs the CLI and Python wheel tarballs produced by the Vercel
+          # preview deployment, but the workflow now starts on pull_request so
+          # tests can begin before the deployment-ready repository_dispatch.
+          # Vercel already reports successful deployments back to GitHub via
+          # deployment statuses and commit statuses. We poll those GitHub-owned
+          # records for candidate public URLs, then trust a URL only after the
+          # actual tarball endpoint responds. That final HTTP check is the
+          # readiness gate, so stale/unrelated status URLs are harmless.
           validate_deployment_url() {
             local url="${1%/}"
             if [ -z "$url" ]; then
@@ -497,6 +500,7 @@ jobs:
     needs:
       - setup
       - unit-test
+      - e2e-test
     permissions:
       statuses: write
     steps:
@@ -532,56 +536,6 @@ jobs:
               sha: '${{ github.event.pull_request.head.sha }}',
               state,
               context: 'Summary',
-              target_url: `${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}`,
-              description: state === 'success' ? 'All checks passed' : 'Some checks failed',
-            });
-      - name: Fail the job if Check All failed
-        if: ${{ steps.check-all.outcome == 'failure' }}
-        run: exit 1
-
-  summary-e2e:
-    name: Summary (e2e)
-    runs-on: ubuntu-latest
-    timeout-minutes: 5
-    if: always()
-    needs:
-      - setup
-      - e2e-test
-    permissions:
-      statuses: write
-    steps:
-      - name: Check All
-        id: check-all
-        continue-on-error: true
-        run: |-
-          for status in ${{ join(needs.*.result, ' ') }}
-          do
-            if [ "$status" != "success" ] && [ "$status" != "skipped" ]
-            then
-              echo "Some checks failed"
-              echo "state=failure" >> $GITHUB_OUTPUT
-              exit 1
-            fi
-          done
-          echo "state=success" >> $GITHUB_OUTPUT
-      # The `context` here MUST exactly match the one written by the
-      # `Post pending Summary statuses to PR commit` step in the `setup`
-      # job above so this final write overwrites that pending row
-      # instead of creating a duplicate.
-      - name: Report status to PR commit
-        if: ${{ always() && github.event.pull_request.head.sha }}
-        uses: actions/github-script@v7
-        env:
-          CHECK_STATE: ${{ steps.check-all.outputs.state }}
-        with:
-          script: |
-            const state = process.env.CHECK_STATE === 'success' ? 'success' : 'failure';
-            await github.rest.repos.createCommitStatus({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              sha: '${{ github.event.pull_request.head.sha }}',
-              state,
-              context: 'Summary (e2e)',
               target_url: `${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}`,
               description: state === 'success' ? 'All checks passed' : 'Some checks failed',
             });
